+++ tiles/src/main/java/com/protomaps/basemap/Basemap.java
@@ -119,7 +119,7 @@ public class Basemap extends ForwardingProfile {
 
   @Override
   public String version() {
-    return "4.11.4";
+    return "4.11.5";
   }
 
   @Override