+++ tiles/src/test/java/com/protomaps/basemap/names/ScriptSegmenterTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
-public class ScriptSegmenterTest {
+class ScriptSegmenterTest {
   @Test
   void getCharScript() {
     // Null