+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Main
+------
+- Fix disputed tag on boundaries with labels only on OSM ways [#475]
+
+Tiles 4.11.5
+------
+- Bump to planetiler v0.9.0
+
 Tiles 4.11.4
 ------
 - More code quality refactorings [#468, #464]